[MOD] Updating Training/Testing Splits
