+++ tensorflow/data/kitti_depth/kitti_depth_val.txt
