+++ tensorflow/data/new_splits/based_on_eigen_split/generate_new_split_kitti_discrete.py
