+++ tensorflow/data/new_splits/based_on_eigen_split/kitti_depth/eigen_train_files.txt
