+++ tensorflow/data/nyudepth/nyudepth_test.txt
