+++ tensorflow/data/nyudepth/nyudepth_train.txt
