+++ tensorflow/data/unreliable_splits/kitti_discrete/kitti_discrete_campus_train.txt
