+++ tensorflow/data/unreliable_splits/kitti_discrete/kitti_discrete_city_train.txt
