+++ tensorflow/modules/datasets/kitti_depth.py
